+++ packages/imandrax-api-models/src/imandrax_api_models/proto_models/task.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from enum import Enum
 
+from imandrax_api import bindings as xbinding
 from pydantic import Field
 
 from ..proto_utils import BaseModel
@@ -22,6 +23,12 @@ class TaskID(BaseModel):
 class Task(BaseModel):
     id: TaskID | None = Field(default=None)
     kind: TaskKind
+
+    def to_proto(self) -> xbinding.task_pb2.Task:
+        task_id = None
+        if self.id is not None:
+            task_id = xbinding.task_pb2.TaskID(id=self.id.id)
+        return xbinding.task_pb2.Task(id=task_id, kind=self.kind.value)
 
 
 class Origin(BaseModel):